requirements.txt
numpy==1.22.3
pdf2image==1.16.0
Pillow==9.0.1
PyPDF2==1.26.0
